Changed the fitting algorithm completely.
Now, instead of doing a binary search, we just count bytes to find the longest string that could possibly fit. This may seem weird, but this is essentially what happens every time we generate a JSON string anyhow. This replaces several iterations in a binary search with a single iteration.